DPRO-3310: Reorganize theme override instructions
There were two things I wanted to improve here.

First, I thought it made more sense to put the journal.yaml configuration
after the introduction to theme overrides, because journal.yaml is a
specific case of the general concept.

Second, there was a factual error where it described Wombat's built-in
theme as "starting from `src/main/webapp/WEB-INF/themes`", which is
actually the location of three themes. I clarified this by listing the
three themes individually.

The three built-in themes made it more complicated to explain the general
case by the email.yaml example. I deleted it and, in its stead, added more
detail to the examples in the following two sections.